style(prettier): add configuration to fix(wallet/ledger): * use catch instead of try catch block * error handling for windows fix(npm): dependency
