+++ package.json
@@ -3,29 +3,17 @@
   "engines": {
     "node": "12.x"
   },
-  "version": "3.6.2",
+  "version": "3.7.0",
   "description": "A desktop application to interact with Terra Core.",
   "author": "Terra <engineering@terra.money> (https://terra.money)",
+  "license": "ISC",
   "dependencies": {
     "@ledgerhq/hw-transport-webhid": "^5.15.0",
     "@ledgerhq/hw-transport-webusb": "^5.15.0",
     "@sentry/browser": "^5.13.0",
-    "@terra-money/use-station": "^3.6.28",
-    "@types/bech32": "^1.1.2",
-    "@types/chart.js": "^2.9.15",
-    "@types/classnames": "^2.2.9",
-    "@types/jest": "^25.1.3",
-    "@types/lodash": "^4.14.149",
-    "@types/node": "^13.7.7",
-    "@types/ramda": "^0.26.43",
-    "@types/react": "^16.9.23",
-    "@types/react-copy-to-clipboard": "^4.3.0",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-modal": "^3.10.5",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/semver": "^7.1.0",
-    "@types/ungap__url-search-params": "^0.1.0",
-    "@ungap/url-search-params": "^0.1.4",
+    "@terra-money/ledger-terra-js": "^1.0",
+    "@terra-money/use-station": "^3.7",
+    "@ungap/url-search-params": "^0.2",
     "axios": "^0.19.2",
     "bech32": "^1.1.3",
     "chart.js": "^2.9.3",
@@ -33,15 +21,15 @@
     "core-js": "^3.6.4",
     "crypto-js": "^4.0.0",
     "deepmerge": "^4.2.2",
+    "i18next": "^19.4.5",
     "intersection-observer": "^0.7.0",
-    "ledger-cosmos-js": "^2.1.7",
     "lodash": "^4.17.15",
-    "node-sass": "^4.13.1",
     "ramda": "^0.27.0",
     "react": "^16.13.0",
     "react-app-polyfill": "^1.0.6",
     "react-copy-to-clipboard": "^5.0.2",
     "react-dom": "^16.13.0",
+    "react-i18next": "^11",
     "react-intersection-observer": "^8.26.1",
     "react-modal": "^3.11.2",
     "react-router-dom": "^5.1.2",
@@ -49,7 +37,6 @@
     "react-toastify": "^5.5.0",
     "secp256k1": "3.7.1",
     "semver": "^7.1.3",
-    "source-map-explorer": "^2.3.1",
     "typescript": "^3.8.3"
   },
   "devDependencies": {
@@ -62,9 +49,26 @@
     "@storybook/addon-viewport": "^5.3.14",
     "@storybook/addons": "^5.3.14",
     "@storybook/react": "^5.3.14",
+    "@types/bech32": "^1.1.2",
+    "@types/chart.js": "^2.9.15",
+    "@types/classnames": "^2.2.9",
+    "@types/lodash": "^4.14.149",
+    "@types/node": "^13.7.7",
+    "@types/ramda": "^0.27",
+    "@types/react": "^16.9.23",
+    "@types/react-copy-to-clipboard": "^4.3.0",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-modal": "^3.10.5",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/semver": "^7.1.0",
+    "babel-loader": "^8",
     "husky": "^4.2.3",
     "lint-staged": "^10.0.8",
-    "prettier": "^1.19.1"
+    "node-sass": "^4.14.1",
+    "prettier": "^1.19.1",
+    "regenerator-runtime": "^0.13",
+    "source-map-explorer": "^2.3.1",
+    "webpack": "4.42.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -83,9 +87,13 @@
       "pre-commit": "lint-staged"
     }
   },
+  "prettier": {
+    "singleQuote": true,
+    "semi": false
+  },
   "lint-staged": {
     "src/**/*.{ts,tsx,json,scss}": [
-      "prettier --single-quote --no-semi --write"
+      "prettier --write"
     ]
   },
   "browserslist": {
