+++ src/post/Confirmation.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, Fragment } from 'react'
 import c from 'classnames'
 import { ConfirmProps, format } from '@terra-money/use-station'
 import { useConfirm, useAuth } from '@terra-money/use-station'
-import getSigner from '../cosmos/signer'
-import signTx from '../cosmos/api/signTx'
+import getSigner from '../wallet/signer'
+import signTx from '../wallet/api/signTx'
 import ModalContent from '../components/ModalContent'
 import Form from '../components/Form'
 import Select from '../components/Select'
