+++ src/wallet/signer.js
